--- mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/touchpoint/domain/model/carousel/CardCarouselItemType.kt
@@ -1,6 +1,0 @@
-package com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domain.model.carousel
-
-enum class CardCarouselItemType() {
-    DEFAULT,
-    FULL;
-}
