+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/touchpoint/view/carousel/CardCarouselItemType.kt
@@ -1,0 +1,6 @@
+package com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.view.carousel
+
+enum class CardCarouselItemType() {
+    DEFAULT,
+    FULL;
+}
