--- mlbusinesscomponents/src/main/res/drawable-hdpi/touchpoint_carousel_card_full_view_background_image.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@color/ui_transparent"/>
-    <corners android:radius="@dimen/touchpoint_card_full_view_radius_border"/>
-</shape>
